+++ src/core/persistence.rs
@@ -43,59 +43,12 @@ pub fn get_version(db: &Connection) -> Result<i32> {
     Ok(version)
 }
 
-pub fn upsert_sensor(db: &Connection, sensor: &Sensor, timestamp: i64) -> Result<()> {
-    db.prepare_cached(
-        // language=sql
-        r#"
-            -- noinspection SqlResolve @ any/"excluded"
-            -- noinspection SqlInsertValues
-            INSERT INTO sensors (sensor_id, title, timestamp, room_title) VALUES (?1, ?2, ?3, ?4)
-            ON CONFLICT (sensor_id) DO UPDATE SET
-                timestamp = excluded.timestamp,
-                title = excluded.title,
-                room_title = excluded.room_title
-        "#,
-    )?
-    .execute(params![sensor.sensor_id, sensor.title, timestamp, sensor.room_title])?;
-    Ok(())
-}
-
-pub fn upsert_reading(db: &Connection, sensor: &Sensor, reading: &Reading) -> Result<()> {
-    let timestamp = reading.timestamp.timestamp_millis();
-    upsert_sensor(&db, &sensor, timestamp)?;
-
-    let mut serialized_value = Vec::new();
-    reading.value.serialize(
-        &mut rmp_serde::Serializer::new(&mut serialized_value)
-            .with_string_variants()
-            .with_struct_map(),
-    )?;
-    db.prepare_cached(
-        // language=sql
-        r#"
-            -- noinspection SqlResolve @ any/"excluded"
-            INSERT INTO readings (sensor_fk, timestamp, value)
-            VALUES ((
-                SELECT pk
-                FROM sensors
-                WHERE sensor_id = ?1
-            ), ?2, ?3)
-            ON CONFLICT (sensor_fk, timestamp) DO UPDATE SET value = excluded.value
-        "#,
-    )?
-    .execute(params![sensor.sensor_id, timestamp, serialized_value])?;
-
-    Ok(())
-}
-
 pub fn select_actuals(db: &Connection) -> Result<Vec<Actual>> {
     db.prepare_cached(
         // language=sql
         r#"
-            SELECT sensors.sensor_id, sensors.title, sensors.timestamp, sensors.room_title, value
+            SELECT sensor_id, title, timestamp, room_title, value
             FROM sensors
-            INNER JOIN readings
-                ON readings.timestamp = sensors.timestamp AND readings.sensor_fk = sensors.pk
             ORDER BY sensors.room_title, sensors.sensor_id
         "#,
     )?
@@ -106,7 +59,7 @@ pub fn select_actuals(db: &Connection) -> Result<Vec<Actual>> {
                 title: row.get("title")?,
                 room_title: row.get("room_title")?,
             },
-            reading: reading_from_row(row)?,
+            reading: get_reading(row)?,
         })
     })?
     .map(|r| r.map_err(Error::from))
@@ -123,7 +76,7 @@ pub fn select_size(db: &Connection) -> Result<u64> {
             SELECT page_count * page_size as size FROM pragma_page_count(), pragma_page_size()
             "#,
         )?
-        .query_row(params![], |row| row.get::<_, i64>(0))
+        .query_row(params![], get_i64)
         .map(|v| v as u64)?)
 }
 
@@ -133,13 +86,12 @@ pub fn select_last_reading(db: &Connection, sensor_id: &str) -> Result<Option<Re
         // language=sql
         .prepare_cached(
             r#"
-            SELECT readings.timestamp, value
+            SELECT timestamp, value
             FROM sensors
-            INNER JOIN readings ON readings.timestamp = sensors.timestamp AND readings.sensor_fk = sensors.pk
             WHERE sensors.sensor_id = ?1
             "#,
         )?
-        .query_row(params![sensor_id], reading_from_row)
+        .query_row(params![sensor_id], get_reading)
         .optional()?)
 }
 
@@ -157,18 +109,79 @@ pub fn select_readings(db: &Connection, sensor_id: &str, since: &DateTime<Local>
             ORDER BY readings.timestamp
             "#,
         )?
-        .query_map(params![sensor_id, since.timestamp_millis()], reading_from_row)?
+        .query_map(params![sensor_id, since.timestamp_millis()], get_reading)?
         .map(|r| r.map_err(Error::from))
         .collect()
 }
 
-fn reading_from_row(row: &Row) -> rusqlite::Result<Reading> {
+fn get_reading(row: &Row) -> rusqlite::Result<Reading> {
     Ok(Reading {
         timestamp: Local.timestamp_millis(row.get("timestamp")?),
         value: rmp_serde::from_read_ref(&row.get::<_, Vec<u8>>("value")?).map_err(|error| {
             rusqlite::Error::FromSqlConversionFailure(0, rusqlite::types::Type::Blob, Box::new(error))
         })?,
     })
+}
+
+fn get_i64(row: &Row) -> rusqlite::Result<i64> {
+    row.get::<_, i64>(0)
+}
+
+impl Value {
+    fn serialize_to_vec(&self) -> Result<Vec<u8>> {
+        let mut serialized_value = Vec::new();
+        self.serialize(
+            &mut rmp_serde::Serializer::new(&mut serialized_value)
+                .with_string_variants()
+                .with_struct_map(),
+        )?;
+        Ok(serialized_value)
+    }
+}
+
+impl Message {
+    pub fn upsert_into(&self, db: &Connection) -> Result<()> {
+        let timestamp = self.reading.timestamp.timestamp_millis();
+        let value = self.reading.value.serialize_to_vec()?;
+
+        db.prepare_cached(
+            // language=sql
+            r#"
+            -- noinspection SqlResolve @ any/"excluded"
+            -- noinspection SqlInsertValues
+            INSERT INTO sensors (sensor_id, title, timestamp, room_title, value) VALUES (?1, ?2, ?3, ?4, ?5)
+            ON CONFLICT (sensor_id) DO UPDATE SET
+                timestamp = excluded.timestamp,
+                title = excluded.title,
+                room_title = excluded.room_title,
+                value = excluded.value
+            "#,
+        )?
+        .execute(params![
+            self.sensor.sensor_id,
+            self.sensor.title,
+            timestamp,
+            self.sensor.room_title,
+            value,
+        ])?;
+
+        db.prepare_cached(
+            // language=sql
+            r#"
+            -- noinspection SqlResolve @ any/"excluded"
+            INSERT INTO readings (sensor_fk, timestamp, value)
+            VALUES ((
+                SELECT pk
+                FROM sensors
+                WHERE sensor_id = ?1
+            ), ?2, ?3)
+            ON CONFLICT (sensor_fk, timestamp) DO UPDATE SET value = excluded.value
+            "#,
+        )?
+        .execute(params![self.sensor.sensor_id, timestamp, value])?;
+
+        Ok(())
+    }
 }
 
 #[cfg(test)]
@@ -178,37 +191,20 @@ mod tests {
     type Result = crate::Result<()>;
 
     #[test]
-    fn double_upsert_sensor_keeps_one_record() -> Result {
-        let sensor = Sensor {
-            sensor_id: "test".into(),
-            title: None,
-            room_title: None,
-        };
-
-        let db = connect(":memory:")?;
-        upsert_sensor(&db, &sensor, 0)?;
-        let sensor_pk = get_sensor_pk(&db, "test")?.unwrap();
-        upsert_sensor(&db, &sensor, 0)?;
-        assert_eq!(get_sensor_pk(&db, "test")?.unwrap(), sensor_pk);
-
-        Ok(())
-    }
-
-    #[test]
-    fn double_upsert_reading_keeps_one_record() -> Result {
+    fn double_upsert_keeps_one_reading() -> Result {
         let message = Composer::new("test")
             .value(Value::Counter(42))
             .timestamp(Local.timestamp_millis(1_566_424_128_000))
             .message;
 
         let db = connect(":memory:")?;
-        upsert_reading(&db, &message.sensor, &message.reading)?;
-        upsert_reading(&db, &message.sensor, &message.reading)?;
+        message.upsert_into(&db)?;
+        message.upsert_into(&db)?;
 
         let reading_count: i64 = db
             // language=sql
             .prepare("SELECT COUNT(*) FROM readings")?
-            .query_row(params![], |row| row.get(0))?;
+            .query_row(params![], get_i64)?;
         assert_eq!(reading_count, 1);
 
         Ok(())
@@ -228,7 +224,7 @@ mod tests {
             .timestamp(Local.timestamp_millis(1_566_424_128_000))
             .message;
         let db = connect(":memory:")?;
-        upsert_reading(&db, &message.sensor, &message.reading)?;
+        message.upsert_into(&db)?;
         assert_eq!(select_last_reading(&db, "test")?, Some(message.reading));
         Ok(())
     }
@@ -240,12 +236,12 @@ mod tests {
             .value(Value::Counter(42))
             .timestamp(Local.timestamp_millis(1_566_424_127_000))
             .message;
-        upsert_reading(&db, &old.sensor, &old.reading)?;
+        old.upsert_into(&db)?;
         let new = Composer::new("test")
             .value(Value::Counter(42))
             .timestamp(Local.timestamp_millis(1_566_424_128_000))
             .message;
-        upsert_reading(&db, &new.sensor, &new.reading)?;
+        new.upsert_into(&db)?;
         assert_eq!(select_last_reading(&db, "test")?, Some(new.reading));
         Ok(())
     }
@@ -257,7 +253,7 @@ mod tests {
             .timestamp(Local.timestamp_millis(1_566_424_128_000))
             .compose();
         let db = connect(":memory:")?;
-        upsert_reading(&db, &message.sensor, &message.reading)?;
+        message.upsert_into(&db)?;
         assert_eq!(
             select_actuals(&db)?,
             vec![Actual {
@@ -275,17 +271,17 @@ mod tests {
             .value(Value::Counter(42))
             .timestamp(Local.timestamp_millis(1_566_424_128_000))
             .message;
-        upsert_reading(&db, &old.sensor, &old.reading)?;
+        old.upsert_into(&db)?;
         let new = Composer::new("test")
             .value(Value::Counter(42))
             .timestamp(Local.timestamp_millis(1_566_424_129_000))
             .message;
-        upsert_reading(&db, &new.sensor, &new.reading)?;
+        new.upsert_into(&db)?;
 
         let reading_count = db
             // language=sql
             .prepare_cached("SELECT COUNT(*) FROM sensors")?
-            .query_row(params![], |row| row.get::<_, i64>(0))?;
+            .query_row(params![], get_i64)?;
         assert_eq!(reading_count, 1);
 
         Ok(())
@@ -299,11 +295,12 @@ mod tests {
         Ok(())
     }
 
+    #[allow(dead_code)]
     fn get_sensor_pk(db: &Connection, sensor_id: &str) -> crate::Result<Option<i64>> {
         Ok(db
             // language=sql
             .prepare_cached("SELECT pk FROM sensors WHERE sensor_id = ?1")?
-            .query_row(params![sensor_id], |row| row.get::<_, i64>(0))
+            .query_row(params![sensor_id], get_i64)
             .optional()?)
     }
 }